+++ src/grid/model/dictionary.interface.ts
@@ -1,0 +1,4 @@
+export interface Dictionary {
+  value: string;
+  display: string;
+}