+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.1.0](https://ri-source.hci.utah.edu/hcidevs/hci-ng-grid/tags/v1.1.0)
     - Configuration added
+    - Global grid service
 
 ## [1.0.0](https://ri-source.hci.utah.edu/hcidevs/hci-ng-grid/tags/v1.0.0)
     - Initial Release.  See README.md for features.